feat(nox-ecs): Add built-in SystemGlobals Archetype (#628)
Currently, it just has tick and time step, but can be expanded to
container other "managed" components in the future such as seed.

Since time step is now available via ECS, it doesn't have to be provided
to `six_dof(...)` by the user. So, it was made optional. The reason for
not removing it altogether is the simulation time step may match the
6DOF time step when running 6DOF in an inner loop at a higher frequency.

Eventually, this can be addressed as part of ELO-336.

Signed-off-by: Akhil Velagapudi <4@4khil.com>